Enable slirp to support user network backend
We used to support the network backend till some time ago (and we used
it in the gnome-boxes snap), but it got broken recently as per missing
support for the user network (when using -nic user,model=virtio arg).

So add it back, by adding the dependency and explicitly enabling such
feature at compile time.

Thanks to https://stackoverflow.com/a/75644037